+++ src/state.rs
@@ -3,10 +3,10 @@ use std::collections::HashMap;
 // State to track previous time step values for each channel
 #[derive(Debug)]
 pub struct RoutingState {
-    pub qup: f64,
-    pub quc: f64,
-    pub qdp: f64,
-    pub depth_p: f64,
+    pub qup: f32,
+    pub quc: f32,
+    pub qdp: f32,
+    pub depth_p: f32,
 }
 
 impl RoutingState {
@@ -19,7 +19,7 @@ impl RoutingState {
         }
     }
 
-    pub fn update(&mut self, quc: f64, qdc: f64, depthc: f64) {
+    pub fn update(&mut self, quc: f32, qdc: f32, depthc: f32) {
         self.qup = self.quc;
         self.quc = quc;
         self.qdp = qdc;
@@ -31,8 +31,8 @@ impl RoutingState {
 #[derive(Debug)]
 pub struct NetworkState {
     pub states: HashMap<String, RoutingState>,
-    pub current_flows: HashMap<String, f64>,
-    pub external_flows: HashMap<String, HashMap<usize, f64>>,
+    pub current_flows: HashMap<String, f32>,
+    pub external_flows: HashMap<String, HashMap<usize, f32>>,
 }
 
 impl NetworkState {
@@ -50,14 +50,14 @@ impl NetworkState {
         self.current_flows.insert(nexus_id.to_string(), 0.0);
     }
 
-    pub fn get_upstream_flow(&self, _nexus_id: &str, upstream_nexuses: &[String]) -> f64 {
+    pub fn get_upstream_flow(&self, _nexus_id: &str, upstream_nexuses: &[String]) -> f32 {
         upstream_nexuses
             .iter()
             .map(|upstream| self.current_flows.get(upstream).unwrap_or(&0.0))
             .sum()
     }
 
-    pub fn update_flow(&mut self, nexus_id: &str, flow: f64) {
+    pub fn update_flow(&mut self, nexus_id: &str, flow: f32) {
         self.current_flows.insert(nexus_id.to_string(), flow);
     }
 }
